Add procedure
To add column time from Orders of Fidelity Activity Trader Pro into Activity and Orders from the Fidelity Website to get a complete report with Time and Date which is needed for further research about Trader behavior. Excel is excellent tool for this and easy then start everything from scratch.